% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil.R
\name{get_soil}
\alias{get_soil}
\title{Retrieve and predict soil samples}
\usage{
get_soil(
  env.id = NULL,
  lat = NULL,
  long = NULL,
  max.depth = NULL,
  isric.data = NULL
)
}
\arguments{
\item{env.id}{character. Identification of the site/environment (e.g. Rayne). The env_id must have length equals 1.}

\item{lat}{numeric. Latitude values of the site/environment (e.g. 33.65) in WGS84. The lat must numeric, length == 1, and between -90 and 90.}

\item{long}{numeric. Longitude values site/environment (e.g. -90.51) in WGS84. The long must numeric, length == 1, and between -180 and 180.}

\item{max.depth}{numeric. The lowest depth soil layer to be consider, in cm (e.g. 20). The depth must numeric, length == 1, and between 5 and 160.}

\item{isric.data}{list. The ISRIC Soil Data Hub (https://data.isric.org) dataset, last access: 27 Oct 2022).}
}
\value{
data.frame The output is a data.frame with the target location and its lat and long coordinates.
Also, the number of samples used, the RMSE, and R-square for each prediction.
The chemicals characteristics are:
CECPH7	Cation exchange capacity - buffered at pH7 cmol(c)/kg
ECEC	Effective cation exchange capacity cmol(c)/kg
NITKJD	Total nitrogen (N) g/kg
ORGC	Organic carbon g/kg
PHAQ	pH H2O unitless
PHPTOT	Phosphorus (P) - total mg/kg
TCEQ	Calcium carbonate equivalent total g/kg
TOTC	Total carbon (C)
The physical characteristics are:
CLAY	Clay total g/100g
SAND	Sand total g/100g
SILT	Silt total g/100g
WG0006	Water retention gravimetric - 6 kPa g/100g
WV0006	Water retention volumetric - 6 kPa cm³/100cm³
}
\description{
This function retrieves soil samples around the world that are near to your target location. Then, predict the soil characteristics for it.
If the number of samples is more than 5, then the function will use Random Forest to make predictions. Otherwise, average the information.
}
\examples{
get_soil(env.id = "RRS", lat = 30.243208 , long = -92.353191, max.depth = 20)
}
