% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil_plot_level.R
\name{get_soil_plot_level}
\alias{get_soil_plot_level}
\title{Predict  soil characteristics at plot level}
\usage{
get_soil_plot_level(samples, plot.polygons)
}
\arguments{
\item{samples}{data.frame The first columns must be sample_id, lat, and long.
The other columns are the soil traits to be extrapolated/predicted at plot level.
The minimum number of samples is 5.}

\item{plot.polygons}{Shapefile/polygons with the id, lat, and long plots information. The coordinate system must be the same for samples and the shapefile.}
}
\value{
data.frame The output is a data.frame with the plot IDs, their lat and long coordinates, the number of samples used to make predictions, the RMSE and R-square for each prediction, the trait predicted, and the predicted value.
}
\description{
This function, from a couple of samples in a trial, predict  soil characteristics at plot level via Random Forest.
}
\examples{
loading some soil samples
samples <- data.table::fread(system.file("soil_data", "soil_samples.txt", package = "SoilType"))
head(samples)
loading the respectively shapefile
library(raster)
plot.polygons <- shapefile(system.file("soil_data", "PI_R1_DAS.shp", package = "SoilType"))
length(plot.polygons) #number of features (plots)
head(plot.polygons@data, 6)
plot.polygons[1,]@bbox
Running the function
test3 <- get_soil_plot_level(samples = samples, plot.polygons = plot.polygons)
and take a look at the results and the average KPI per soil trait
head(test3)
tapply(test3$Rsquared, test3$Trait, mean)
}
